Update to Vite v4
Update to newest version of Vite v4.

There didn't seem to be any configuration changes needed for this to produce a bundle.

Behind the scenes, the biggest change here is probably the upgrade to Rollup v3. There is also another bump in the browser baseline.

For more details, see the migration guide at: https://vitejs.dev/guide/migration.html